Update consumeEvents method in OpenbookProgram
The consumeEvents method in OpenbookProgram has been updated to include 'openOrdersAccounts' as a new parameter. This enhancement allows for the generation of keys from the public keys in the 'openOrdersAccounts' list. The changes are reflected in the unit test in OpenBookTest.java, where the handling of OpenBookMarkets was adjusted to accommodate the new parameter in consumeEvents.